Refactor: Imported IRepository to ProductRepository
