--- HenriksHobbyLager/ProductRepository.cs
@@ -1,25 +1,0 @@
-﻿using RefactoringExercise;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-
-namespace HenriksHobbyLager
-{
-    internal class ProductRepository : IRepository<T>
-    {
-        public class Product
-        {
-            public int Id { get; set; }
-            public string Name { get; set; }
-            public decimal Price { get; set; }
-            public int Stock { get; set; }
-            public string Category { get; set; }
-            public DateTime Created { get; set; }
-            public DateTime? LastUpdated { get; set; }  // Frågetecknet är för att jag är osäker på datumet
-        }
-
-    }
-}
